feat: add lite bundle and code split
